% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{CMP Normalizing Constant}
\alias{CMP Normalizing Constant}
\alias{ncmp}
\title{TBD: Update this description
TBD: Pass around options instead of relying on defaults.
Brute force computation of the z-function. This computation is done at the
original scale (as opposed to the log-scale), so it becomes unstable when
the magnitudes of the terms become very large.
Sum (j>=0) { lambda^j / (j!)^nu }}
\usage{
ncmp(lambda, nu, log = FALSE, method = "hybrid")
}
\description{
TBD: Update this description
TBD: Pass around options instead of relying on defaults.
Brute force computation of the z-function. This computation is done at the
original scale (as opposed to the log-scale), so it becomes unstable when
the magnitudes of the terms become very large.
Sum (j>=0) { lambda^j / (j!)^nu }
}
