\name{couple.rda}
\docType{data}
\alias{couple}
\title{Couple dataset}
\description{
A dataset investigating the impact of education level and level of anxious
attachment on unwanted pursuit behaviors in the context of couple separation.
}
\usage{
data(couple)
}
\format{
\itemize{
\item UPB = number of unwanted pursuit behavior perpetrations.
\item EDUCATION = 1 if at least bachelor’s degree; 0 otherwise.
\item ANXIETY = continuous measure of anxious attachment.
}
}

\references{
Loeys, T., Moerkerke, B., DeSmet, O., Buysse, A., 2012. The analysis of zero-inflated count data: Beyond zero-inflated Poisson regression. British J. Math. Statist. Psych. 65 (1), 163-180.
}
