% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freight}
\alias{freight}
\title{Freight dataset}
\format{
\describe{
\item{broken}{number of ampules found broken upon arrival.}
\item{transfers}{number of times carton was transferred from one aircraft to
another.}
}
}
\usage{
data(freight)
}
\description{
A set of data on airfreight breakage (breakage of ampules filled with some
biological substance are shipped in cartons).
}
\references{
Kutner MH, Nachtsheim CJ, Neter J (2003). Applied Linear Regression Models,
Fourth Edition. McGraw-Hill.
}
\keyword{datasets}
