\name{freight.rda}
\docType{data}
\alias{freight}
\title{Freight dataset}
\description{
  A set of data on airfreight breakage (breakage of ampules filled with some biological substance are shipped in cartons).
}
\usage{
data(freight)
}
\format{
broken = number of ampules found broken upon arrival.

transfers = number of times carton was transferred from one aircraft to another
}

\references{
	Kutner MH, Nachtsheim CJ, Neter J (2003). Applied Linear Regression Models, Fourth Edition. McGraw-Hill.
}