% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.init}
\alias{get.init}
\title{Construct initial values for coefficients.}
\usage{
get.init(beta = NULL, gamma = NULL, zeta = NULL)
}
\arguments{
\item{beta}{Vector for \code{beta}.}

\item{gamma}{Vector for \code{gamma}.}

\item{zeta}{Vector for \code{zeta}.}
}
\value{
List of initial value terms.
}
\description{
Construct initial values for coefficients.
}
\details{
The default value \code{NULL} is interpreted as an empty vector, so that the
given component is absent from the model.
}
