% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.init.zero}
\alias{get.init.zero}
\title{Construct initial values for coefficients with zeros.}
\usage{
get.init.zero(d1 = 0, d2 = 0, d3 = 0)
}
\arguments{
\item{d1}{Dimension of \code{beta}.}

\item{d2}{Dimension of \code{gamma}.}

\item{d3}{Dimension of \code{zeta}.}
}
\value{
List of initial value terms.
}
\description{
Construct initial values for coefficients with zeros.
}
