% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.modelmatrix}
\alias{get.modelmatrix}
\title{Construct model matrices and offsets for CMP/ZICMP regression}
\usage{
get.modelmatrix(X = NULL, S = NULL, W = NULL, offset = NULL)
}
\arguments{
\item{X}{An \code{X} matrix to use with \code{beta}.}

\item{S}{An \code{S} matrix to use with \code{gamma}.}

\item{W}{A \code{W} matrix to use with \code{zeta}.}

\item{offset}{An offset object constructed from \code{get.offset}.}
}
\value{
List of model matrix terms.
}
\description{
Construct model matrices and offsets for CMP/ZICMP regression
}
