% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.offset}
\alias{get.offset}
\title{Construct values for offsets.}
\usage{
get.offset(x = NULL, s = NULL, w = NULL)
}
\arguments{
\item{x}{Vector of offsets to go with \code{X} matrix.}

\item{s}{Vector of offsets to go with \code{S} matrix.}

\item{w}{Vector of offsets to go with \code{W} matrix.}
}
\value{
List of offset terms.
}
\description{
Construct values for offsets.
}
\details{
The default value \code{NULL} is interpreted as a vector of zeros. At least
one component must be non-NULL so that the dimension can be determined.
}
