% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip-reg.R
\name{glm.cmp, ZIP support}
\alias{glm.cmp, ZIP support}
\alias{summary.zip}
\alias{print.zip}
\alias{logLik.zip}
\alias{AIC.zip}
\alias{BIC.zip}
\alias{coef.zip}
\alias{sdev.zip}
\alias{vcov.zip}
\alias{deviance.zip}
\alias{residuals.zip}
\alias{predict.zip}
\alias{parametric.bootstrap.zip}
\title{Supporting Functions for ZIP Regression}
\usage{
\method{summary}{zip}(object, ...)

\method{print}{zip}(x, ...)

\method{logLik}{zip}(object, ...)

\method{AIC}{zip}(object, ..., k = 2)

\method{BIC}{zip}(object, ...)

\method{coef}{zip}(object, type = c("vector", "list"), ...)

\method{sdev}{zip}(object, type = c("vector", "list"), ...)

\method{vcov}{zip}(object, ...)

\method{deviance}{zip}(object, ...)

\method{residuals}{zip}(object, type = c("raw", "quantile"), ...)

\method{predict}{zip}(object, newdata = NULL, type = c("response", "link"), ...)

\method{parametric.bootstrap}{zip}(object, reps = 1000, report.period = reps + 1, ...)
}
\arguments{
\item{object}{object of type \code{zip}.}

\item{...}{other arguments, such as \code{subset} and \code{na.action}.}

\item{x}{object of type \code{zip}.}

\item{k}{Penalty per parameter to be used in AIC calculation.}

\item{type}{Specifies quantity to be computed. See details.}

\item{newdata}{New covariates to be used for prediction.}

\item{reps}{Number of bootstrap repetitions.}

\item{report.period}{Report progress every \code{report.period} iterations.}
}
\description{
Supporting Functions for ZIP Regression
}
\details{
The function \code{residuals} returns raw residuals when
\code{type = "raw"}  and quantile residuals when
\code{type = "quantile"}.

The function \code{predict} returns expected values of the outcomes,
eveluated at the computed estimates, when \code{type = "response"}. When
\code{type = "link"}, a \code{data.frame} is instead returned with
columns corresponding to estimates of \code{lambda}, \code{nu}, and
\code{p}.

The function \code{coef} returns a vector of coefficient estimates in
the form \code{c(beta, zeta)} when \code{type = "vector"}. When
\code{type = "list"}, the estimates are returned as a list with named
elements \code{beta} and \code{zeta}.

The \code{type} argument behaves the same for the \code{sdev} function
as it does for \code{coef}.
}
