% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{parametric.bootstrap}
\alias{parametric.bootstrap}
\title{Parametric Bootstrap}
\usage{
parametric.bootstrap(object, reps = 1000, report.period = reps + 1,
  ...)
}
\arguments{
\item{object}{a model object}

\item{reps}{Number of bootstrap repetitions.}

\item{report.period}{Report progress every \code{report.period} iterations.}

\item{...}{other parameters which might be required by the model}
}
\value{
The form of the value returned depends on the class of its argument. See
the documentation of the particular methods for details of what is produced by that method.
}
\description{
A generic function for the parametric bootstrap from the results of
various model fitting functions. The function invokes particular methods
which depend on the class of the first argument.
}
\details{
See the documentation of the particular methods for details.
}
\author{
Thomas Lotze
}
