% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{CMP Distribution}
\alias{CMP Distribution}
\alias{dcmp}
\alias{rcmp}
\alias{pcmp}
\alias{qcmp}
\alias{ecmp}
\alias{vcmp}
\alias{ncmp}
\alias{tcmp}
\title{COM-Poisson Distribution}
\usage{
dcmp(x, lambda, nu, log = FALSE)

rcmp(n, lambda, nu)

pcmp(x, lambda, nu)

qcmp(q, lambda, nu, log.p = FALSE)

ecmp(lambda, nu, method = "hybrid")

vcmp(lambda, nu, method = "hybrid")

ncmp(lambda, nu, log = FALSE, method = "hybrid")

tcmp(lambda, nu)
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda}{rate parameter.}

\item{nu}{dispersion parameter.}

\item{log}{logical; if TRUE, probabilities are returned on log-scale.}

\item{n}{number of observations.}

\item{q}{vector of probabilities.}

\item{log.p}{logical; if TRUE, probabilities \code{p} are given as \eqn{\log(p)}.}

\item{method}{a string: \code{hybrid}, \code{approx}, or \code{hybrid}.}
}
\value{
\describe{
\item{dcmp}{gives the density,}
\item{pcmp}{gives the cumulative probability,}
\item{qcmp}{gives the quantile function,}
\item{rcmp}{generates random values,}
\item{ecmp}{gives the expected value,}
\item{vcmp}{gives the variance,}
\item{ncmp}{gives the value of the normalizing constant, and}
\item{tcmp}{gives the truncating upper value (see details).}
}
}
\description{
Functions for the COM-Poisson distribution.
}
\details{
TBD: tcmp
}
\references{
Kimberly F. Sellers & Galit Shmueli (2010). A Flexible Regression Model for
Count Data. Annals of Applied Statistics, 4(2), 943-961.
}
\author{
Kimberly Sellers
}
\keyword{COM-Poisson}
\keyword{distribution}
