% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normconst.R
\name{CMP Normalizing Constant}
\alias{CMP Normalizing Constant}
\alias{ncmp}
\title{CMP Normalizing Constant}
\usage{
ncmp(lambda, nu, log = FALSE, method = "hybrid")
}
\arguments{
\item{lambda}{rate parameter.}

\item{nu}{dispersion parameter.}

\item{log}{If \code{TRUE}, return value on the log-scale.}

\item{method}{\code{trunc}, \code{approx}, or \code{hybrid}. See details.}
}
\description{
CMP Normalizing Constant
}
\details{
Compute the normalizing constant
\deqn{
Z(\lambda, \nu) = \sum_{x=0}^\infty \lambda^x / (x!)^\nu
}

Available methods are:
\describe{
\item{trunc}{Compute a truncated version of the series \eqn{Z(\lambda, \nu)}.}
\item{approx}{Use the approximation
\deqn{
Z(\lambda, \nu) = \frac{ \exp(\nu \lambda^{1/\nu}) }{ \lambda^{(\nu-1)/2\nu} (
2\pi)^{(\nu-1)/2} \nu^{1/2} }
}
}
\item{hybrid}{Use the \code{approx} method if \eqn{\lambda^{1/\nu}} is small;
otherwise, use the \code{trunc} method. The \code{hybrid} method is used
when computing the density.}
}
}
