% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{COMPoissonReg-options}
\alias{COMPoissonReg-options}
\title{Package options}
\arguments{
\item{COMPoissonReg.control}{A default control data structure for the
package. See the helper function \link{get.control} for a description of
contents.}
}
\description{
Global options used by the COMPoissonReg package.
}
\details{
\itemize{
\item \code{getOption("COMPoissonReg.control")}
}
}
