% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R, R/cmp-reg.R, R/zicmp-reg.R
\name{equitest}
\alias{equitest}
\alias{equitest.cmp}
\alias{equitest.zicmp}
\title{Equidispersion Test}
\usage{
equitest(object, ...)

\method{equitest}{cmp}(object, ...)

\method{equitest}{zicmp}(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{other parameters which might be required by the model}
}
\value{
Returns the test statistic and p-value determined from the \eqn{\chi.1^2}
distribution.
}
\description{
Likelihood ratio test for Equidispersion
}
\details{
A generic function for the likelihood ratio test for
equidispersion using the output of a fitted mode. The function invokes
particular methods which depend on the class of the first argument.
}
\author{
Thomas Lotze
}
