% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{equitest}
\alias{equitest}
\title{Equidispersion Test}
\usage{
equitest(object, ...)
}
\arguments{
\item{object}{a model object}

\item{...}{other parameters which might be required by the model}
}
\value{
Returns the test statistic and p-value determined from a \eqn{\chi^2}
distribution with \eqn{d_2} degrees of freedom.
}
\description{
Likelihood ratio test for equidispersion
}
\details{
A generic function for the likelihood ratio test for
equidispersion using the output of a fitted mode. The function invokes
particular methods which depend on the class of the first argument.
}
\author{
Thomas Lotze
}
