% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.control}
\alias{get.control}
\title{Construct a control object to pass additional arguments to a number of
functions in the package.}
\usage{
get.control(
  ymax = 1e+06,
  z.method = "hybrid",
  optim.method = "L-BFGS-B",
  optim.control = list(maxit = 150),
  grad.eps = 1e-05,
  hess.eps = 0.01,
  hybrid.tol = 0.01,
  truncate.tol = 1e-06
)
}
\arguments{
\item{ymax}{Truncate counts to maximum value of \code{y}. Protects against
very long computations.}

\item{z.method}{One of: \code{"hybrid"}, \code{"approx"}, or \code{"trunc"}.
See details.}

\item{optim.method}{Optimization method for maximum likelihood. See the
\code{method} argument in \link{stats::optim}.}

\item{optim.control}{\code{control} argument for \link{stats::optim}.}

\item{grad.eps}{TBD: remove?}

\item{hess.eps}{TBD: remove?}

\item{hybrid.tol}{Tolerance for \code{z.method = "hybrid"} to use
approximation or truncation approaches to compute normalizing constant and
related quantities. See package vignette for details.}

\item{truncate.tol}{Tolerance for truncation approach to compute normalizing
constant and related quantities.}
}
\value{
List of controls.
}
\description{
Construct a control object to pass additional arguments to a number of
functions in the package.
}
