% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get.offset.zero}
\alias{get.offset.zero}
\title{Construct zero values for offsets.}
\usage{
get.offset.zero(n)
}
\arguments{
\item{n}{Number of observations.}
}
\value{
List of offset terms containing all zeros.
}
\description{
Construct zero values for offsets.
}
