% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.cmp-raw}
\alias{glm.cmp-raw}
\alias{glm.cmp.raw}
\alias{glm.zicmp.raw}
\title{Raw Interface to COM-Poisson and Zero-Inflated COM-Poisson Regression}
\usage{
glm.cmp.raw(y, X, S, offset = NULL, init = NULL, fixed = NULL, control = NULL)

glm.zicmp.raw(
  y,
  X,
  S,
  W,
  offset = NULL,
  init = NULL,
  fixed = NULL,
  control = NULL
)
}
\arguments{
\item{y}{A vector of counts which represent the response .}

\item{X}{Design matrix for the \code{lambda} regression.}

\item{S}{Design matrix for the \code{nu} regression.}

\item{offset}{A data structure that specifies offsets. See the helper
function \link{get.offset}.}

\item{init}{A data structure that specifies initial values. See the helper
function \link{get.init}.}

\item{fixed}{A data structure that specifies which coefficients should
remain fixed in the maximum likelihood procedure. See the helper function
\link{get.fixed}.}

\item{control}{A control data structure. See the helper function
\link{get.control}.}

\item{W}{Design matrix for the \code{p} regression.}
}
\value{
See the \link{glm.cmp}.
}
\description{
Fit COM-Poisson and Zero-Inflated COM-Poisson regression using a "raw"
interface which bypasses the formula-driven interface of \code{glm.cmp}.
}
