% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm.cmp-raw}
\alias{glm.cmp-raw}
\alias{glm.cmp.raw}
\alias{glm.zicmp.raw}
\title{Raw Interface to COM-Poisson and Zero-Inflated COM-Poisson Regression}
\usage{
glm.cmp.raw(y, X, S, offset = NULL, init = NULL, fixed = NULL, control = NULL)

glm.zicmp.raw(
  y,
  X,
  S,
  W,
  offset = NULL,
  init = NULL,
  fixed = NULL,
  control = NULL
)
}
\arguments{
\item{y}{A vector of counts which represent the response .}

\item{X}{Design matrix for the `lambda` regression.}

\item{S}{Design matrix for the `nu` regression.}

\item{W}{Design matrix for the `p` regression.}

\item{off.x}{Offset term for the `lambda` regression. If \code{NULL}, offsets
are taken to be zeros.}

\item{off.s}{Offset term for the `nu` regression. If \code{NULL}, offsets
are taken to be zeros.}

\item{off.w}{Offset term for the `p` regression. If \code{NULL}, offsets
are taken to be zeros.}

\item{beta.init}{Initial value for `beta`. If \code{NULL}, initial value
is taken to be a vector of zeros.}

\item{gamma.init}{Initial value for `gamma`. If \code{NULL}, initial value
is taken to be a vector of zeros.}

\item{zeta.init}{Initial value for `zeta`. If \code{NULL}, initial value
is taken to be a vector of zeros.}
}
\value{
See the \link{glm.cmp}.
}
\description{
Fit COM-Poisson and Zero-Inflated COM-Poisson regression using a "raw"
interface which bypasses the formula-driven interface of \code{glm.cmp}.
}
