% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp-reg.R, R/zicmp-reg.R
\name{glm.cmp, ZICMP support}
\alias{glm.cmp, ZICMP support}
\alias{get.cmp.newdata}
\alias{summary.zicmp}
\alias{print.zicmp}
\alias{logLik.zicmp}
\alias{AIC.zicmp}
\alias{BIC.zicmp}
\alias{coef.zicmp}
\alias{nu.zicmp}
\alias{sdev.zicmp}
\alias{vcov.zicmp}
\alias{deviance.zicmp}
\alias{residuals.zicmp}
\alias{get.zicmp.newdata}
\alias{predict.zicmp}
\alias{parametric.bootstrap.zicmp}
\title{Supporting Functions for ZICMP Regression}
\usage{
get.cmp.newdata(X, S, offset = get.offset())

\method{summary}{zicmp}(object, ...)

\method{print}{zicmp}(x, ...)

\method{logLik}{zicmp}(object, ...)

\method{AIC}{zicmp}(object, ..., k = 2)

\method{BIC}{zicmp}(object, ...)

\method{coef}{zicmp}(object, type = c("vector", "list"), ...)

\method{nu}{zicmp}(object, ...)

\method{sdev}{zicmp}(object, type = c("vector", "list"), ...)

\method{vcov}{zicmp}(object, ...)

\method{deviance}{zicmp}(object, ...)

\method{residuals}{zicmp}(object, type = c("raw", "quantile"), ...)

get.zicmp.newdata(X, S, W, offset)

\method{predict}{zicmp}(object, newdata = NULL, type = c("response", "link"), ...)

\method{parametric.bootstrap}{zicmp}(object, reps = 1000, report.period = reps + 1, ...)
}
\arguments{
\item{object}{object of type \code{zicmp}.}

\item{...}{other arguments, such as \code{subset} and \code{na.action}.}

\item{x}{object of type \code{zicmp}.}

\item{k}{Penalty per parameter to be used in AIC calculation.}

\item{type}{Specifies quantity to be computed. See details.}

\item{newdata}{New covariates to be used for prediction.}

\item{reps}{Number of bootstrap repetitions.}

\item{report.period}{Report progress every \code{report.period} iterations.}
}
\description{
Supporting Functions for ZICMP Regression
}
\details{
The function \code{residuals} returns raw residuals when
\code{type = "raw"}  and quantile residuals when
\code{type = "quantile"}.

The function \code{predict} returns expected values of the outcomes,
eveluated at the computed estimates, when \code{type = "response"}. When
\code{type = "link"}, a \code{data.frame} is instead returned with
columns corresponding to estimates of \code{lambda}, \code{nu}, and
\code{p}.

The function \code{coef} returns a vector of coefficient estimates in
the form \code{c(beta, gamma, zeta)} when \code{type = "vector"}. When
\code{type = "list"}, the estimates are returned as a list with named
elements \code{beta} and \code{gamma}, and \code{zeta}.

The \code{type} argument behaves the same for the \code{sdev} function
as it does for \code{coef}.
}
