% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zicmp.R
\name{ZICMP Distribution}
\alias{ZICMP Distribution}
\alias{dzicmp}
\alias{rzicmp}
\alias{pzicmp}
\alias{qzicmp}
\alias{ezicmp}
\alias{vzicmp}
\title{ZICMP Distribution}
\usage{
dzicmp(x, lambda, nu, p, log = FALSE)

rzicmp(n, lambda, nu, p)

pzicmp(x, lambda, nu, p)

qzicmp(q, lambda, nu, p, log.p = FALSE)

ezicmp(lambda, nu, p, method = "hybrid")

vzicmp(lambda, nu, p, method = "hybrid")
}
\arguments{
\item{x}{vector of quantiles.}

\item{lambda}{rate parameter.}

\item{nu}{dispersion parameter.}

\item{p}{zero-inflation probability parameter.}

\item{log}{logical; if TRUE, probabilities are returned on log-scale.}

\item{n}{number of observations.}

\item{q}{vector of probabilities.}

\item{log.p}{logical; if TRUE, probabilities p are given as \eqn{\log(p)}.}

\item{method}{a string: \code{hybrid}, \code{approx}, or \code{trunc}.}
}
\value{
\describe{
\item{dzicmp}{the density,}
\item{pzicmp}{the cumulative probability,}
\item{qzicmp}{the quantile value, and}
\item{rzicmp}{generates random numbers,}
\item{ezicmp}{the expected value.}
\item{vzicmp}{the variance.}
}
}
\description{
Computes the density, cumulative probability, quantiles, and
random draws for the zero-inflated COM-Poisson distribution.
}
\references{
Kimberly F. Sellers and Andrew M. Raim (2016). A Flexible Zero-Inflated Model
to Address Data Dispersion. Computational Statistics and Data Analysis, 99,
68-80.
}
\author{
Kimberly Sellers, Andrew Raim
}
